+++ jib-core/src/main/java/com/google/cloud/tools/jib/registry/RegistryClient.java
@@ -370,18 +370,17 @@ public class RegistryClient {
   /**
    * Check if an image is on the registry.
    *
-   * @param imageDigest the image digest to check for
-   * @return the image's {@link ManifestAndDigest} if the image exists on the registry, or {@link
-   *     Optional#empty()} if it doesn't
+   * @param imageQualifier the tag or digest to check for
+   * @return the {@link ManifestAndDigest} of the image if the image exists on the registry, or
+   *     {@link Optional#empty()} otherwise
    * @throws IOException if communicating with the endpoint fails
    * @throws RegistryException if communicating with the endpoint fails
    */
-  public Optional<ManifestAndDigest<ManifestTemplate>> checkImage(String imageDigest)
+  public Optional<ManifestAndDigest<ManifestTemplate>> checkImage(String imageQualifier)
       throws IOException, RegistryException {
     ManifestChecker<ManifestTemplate> manifestChecker =
         new ManifestChecker<>(
-            registryEndpointRequestProperties, imageDigest, ManifestTemplate.class);
-
+            registryEndpointRequestProperties, imageQualifier, ManifestTemplate.class);
     return callRegistryEndpoint(manifestChecker);
   }
 
@@ -389,7 +388,7 @@ public class RegistryClient {
    * Pulls the image manifest and digest for a specific tag.
    *
    * @param <T> child type of ManifestTemplate
-   * @param imageTag the tag to pull on
+   * @param imageQualifier the tag or digest to pull on
    * @param manifestTemplateClass the specific version of manifest template to pull, or {@link
    *     ManifestTemplate} to pull predefined subclasses; see: {@link
    *     ManifestPuller#handleResponse(Response)}
@@ -398,14 +397,16 @@ public class RegistryClient {
    * @throws RegistryException if communicating with the endpoint fails
    */
   public <T extends ManifestTemplate> ManifestAndDigest<T> pullManifest(
-      String imageTag, Class<T> manifestTemplateClass) throws IOException, RegistryException {
+      String imageQualifier, Class<T> manifestTemplateClass) throws IOException, RegistryException {
     ManifestPuller<T> manifestPuller =
-        new ManifestPuller<>(registryEndpointRequestProperties, imageTag, manifestTemplateClass);
+        new ManifestPuller<>(
+            registryEndpointRequestProperties, imageQualifier, manifestTemplateClass);
     return callRegistryEndpoint(manifestPuller);
   }
 
-  public ManifestAndDigest<?> pullManifest(String imageTag) throws IOException, RegistryException {
-    return pullManifest(imageTag, ManifestTemplate.class);
+  public ManifestAndDigest<?> pullManifest(String imageQualifier)
+      throws IOException, RegistryException {
+    return pullManifest(imageQualifier, ManifestTemplate.class);
   }
 
   /**
